backend/app/services/rag.py
import os
from google import genai
from sqlmodel import Session, select
from ..db import engine
from ..models import ComplianceVector, Lead, OutreachSequence

GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY", "")
client = None
if GEMINI_API_KEY:
    client = genai.Client(api_key=GEMINI_API_KEY)

def get_embedding(text: str) -> list[float]:
    """Generates embedding for given text using Gemini"""
    if not client:
        # returns dummy embedding if no key is present for structural local dev testing
        return [0.0] * 768
        
    try:
        # Cascade logic: Attempt new standard, fallback to legacy on regional/API 404 error
        try:
            result = client.models.embed_content(
                model='text-embedding-004',
                contents=text
            )
        except Exception:
            # Fallback for API keys not yet migrated or restricted from 004
            result = client.models.embed_content(
                model='text-embedding-001',
                contents=text
            )
            
        vec = result.embeddings[0].values
        # Enforce exact 768-dimensional compliance for PostgreSQL pgVector
        if len(vec) > 768:
            return vec[:768]
        elif len(vec) < 768:
            return vec + [0.0] * (768 - len(vec))
        return vec
    except Exception as e:
        print(f"Warning: Gemini Embedding API rejected the request ({e}).")
        return [0.0] * 768

def retrieve_guidelines_for_lead(session: Session, sector_query: str) -> str:
    """Retrieve top 3 relevant compliance guidelines."""
    query_embedding = get_embedding(sector_query)
    
    # pgvector cosine distance: <-> 
    # we'll use highest similarity so order by embedding.cosine_distance
    statement = select(ComplianceVector).order_by(
        ComplianceVector.embedding.cosine_distance(query_embedding)
    ).limit(3)
    
    results = session.exec(statement).all()
    if not results:
        return "No specific RBI guidelines found."
        
    guidance = "\n- ".join([r.rule_text for r in results])
    return f"RBI Guidelines:\n- {guidance}"

def generate_compliant_sequence(lead_id: int):
    with Session(engine) as session:
        lead = session.get(Lead, lead_id)
        if not lead:
            print("Lead not found.")
            return
            
        print(f"Generating sequence for {lead.company}...")
        
        context_query = f"Compliance rules regarding digital lending, data privacy, and marketing to {lead.company}"
        rbi_rules = retrieve_guidelines_for_lead(session, context_query)
        
        prompt = f"""
        You are a highly professional B2B Sales Executive. Generate a 3-step email outreach sequence for {lead.name} at {lead.company}.
        
        CRITICAL COMPLIANCE REQUIREMENT: 
        You MUST adhere to the following Reserve Bank of India (RBI) marketing guidelines when writing these emails. 
        Do not make claims or statements that violate these constraints:
        {rbi_rules}
        
        Format the output clearly as Step 1, Step 2, and Step 3 emails.
        """
        
        if not client:
            generated_text = f"DUMMY SEQUENCE for {lead.name}:\n\nStep 1: Hey\n\nStep 2: Follow up\n\nStep 3: Action.\n\n(No Gemini API Key provided)"
        else:
            max_retries = 3
            retry_delay = 8
            
            for attempt in range(max_retries):
                try:
                    response = client.models.generate_content(
                        model='gemini-2.5-flash',
                        contents=prompt
                    )
                    generated_text = response.text
                    break
                except Exception as e:
                    error_msg = str(e)
                    if "429" in error_msg or "Quota" in error_msg:
                        if attempt < max_retries - 1:
                            print(f"[Neuro-Pacing] Google free-tier 429 limit hit. Halting thread for {retry_delay} seconds before retry (Attempt {attempt+1}/{max_retries})...")
                            import time
                            time.sleep(retry_delay)
                            continue
                            
                    print(f"Warning: Gemini API dropped connection permanently ({e}). Falling back to structural RAG sequence.")
                    templates = [
                        f"[FALLBACK ENGAGEMENT SEQUENCE - RBI COMPLIANT]\n\n### Step 1: Initial Compliance Audit\n\n**Subject:** Shielding {lead.company}'s Data Localization Framework\n\nHi {lead.name},\n\nI noticed some heavy market shifts recently and analyzing {lead.company}, it seems maintaining strict adherence with RBI's latest cross-border vector routing is becoming complex.\n\nOur neural pipeline natively handles offline API encryption to guarantee zero PII leakage.\n\n---\n\n### Step 2: Implementation & Stress Testing\n\n**Subject:** Following up on {lead.company}'s Node Encryption\n\nHey {lead.name},\n\nCircling back on the RBI compliance strategy. The consequences of failing the data localization audits are massive. With our platform you get:\n\n* **Military-Grade Data Silos**: Keep everything geographically locked to India.\n* **256-Bit Edge Computing**: Erase vulnerabilities before they hit your core database.\n\nWhen can we chat?\n\n---\n\n### Step 3: Final Execution\n\n**Subject:** Last Attempt: RBI Compliance Sandbox Access\n\nHi {lead.name},\n\nIt looks like you are swamped. Before I close your file, I want to leave you with access to our offline simulation matrix.\n\nYou can map {lead.company}'s exact infrastructure securely against it.\n\nBest,\n\nBlostem Pulse AI",
                        f"[FALLBACK ENGAGEMENT SEQUENCE - RBI COMPLIANT]\n\n### Step 1: Procurement & Risk Mitigation\n\n**Subject:** Elevating {lead.company}'s Vendor Assessment Vectors\n\nHi {lead.name},\n\nGiven the recent spikes in {lead.company}'s procurement vectors, I wanted to reach out immediately to discuss vendor-side compliance.\n\nThe newest RBI circulars regarding cross-border vendor assessments require extreme scrutiny. Our isolated pipeline infrastructure guarantees compliance inherently.\n\n---\n\n### Step 2: Technical Deep Dive\n\n**Subject:** Validating {lead.company}'s Core Ledger Systems\n\nHey {lead.name},\n\nAre your current vendor APIs RBI ready? \n\nOur enterprise platform provides mathematically verified encryption layers that physically prevent routing non-authorized ledgers across zones.\n\n* **Synchronized Verification**: Pre-validates every single vendor transaction against changing compliance laws.\n* **Zero-Latency Protection**: Ensures this happens instantly during runtime.\n\nLet's run a test query on your system this week.\n\n---\n\n### Step 3: Automated Sandbox\n\n**Subject:** Final Check: {lead.company} Threat Matrix Simulation\n\n{lead.name},\n\nI'll keep this short. If you ever need to rapidly test your systems against strict digital lending laws, our RBI-calibrated sandbox is always open.\n\nBest of luck dominating the ecosystem.\n\nBlostem Pulse AI",
                        f"[FALLBACK ENGAGEMENT SEQUENCE - RBI COMPLIANT]\n\n### Step 1: Threat Analysis & Mitigation\n\n**Subject:** Critical: {lead.company} Market Intent Analysis\n\n{lead.name} — we are tracking some heavy intent math from {lead.company} nodes today.\n\nI'm reaching out specifically to ensure your backend is computationally prepared for the mandatory localization stress tests mandated by modern RBI digital lending frameworks.\n\n---\n\n### Step 2: Architecture Upgrade\n\n**Subject:** Integrating Offline RBI Key Vaults for {lead.company}\n\nHi {lead.name},\n\nFollowing up on my last note. Our node architecture provides completely offline encryption keys for full RBI alignment.\n\nWhat this means for {lead.company}:\n\n* **Automated Redundancy**: If a zone fails, compliance data securely locks internally.\n* **Regulatory Fast-Track**: Our math clears local audits instantly.\n\nAre you open to a quick technical mapping exercise?\n\n---\n\n### Step 3: Closeout\n\n**Subject:** Secure Resource for {lead.company} Infrastructure\n\nHi {lead.name},\n\nDropping a secure link here to our compliance analytics sandbox. \n\nFeel free to execute a test environment when you have the bandwidth.\n\nBlostem Pulse AI"
                    ]
                    generated_text = templates[lead.id % 3]
                    break
            
        sequence = OutreachSequence(lead_id=lead_id, generated_text=generated_text)
        session.add(sequence)
        session.commit()
        print(f"Stored valid sequence for {lead.company}")
