backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
celery==5.3.6
redis==5.0.3
sqlmodel==0.0.16
psycopg2-binary==2.9.9
pgvector==0.2.5
pydantic>=2.10.0
pydantic-settings==2.2.1
google-genai==0.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
